requirements.txt
discord.py
python-dotenv
requests
pandas
numpy
rapidfuzz
pymongo[srv]